requirements.txt
paho-mqtt
unidecode
python-twitter
